some changes

+++ app/Http/Controllers/Api/V1/SuborganizationController.php
@@ -383,6 +383,7 @@ class SuborganizationController extends Controller
             }
         }
 
+
         $invited = $this->organizationRepository->inviteMember($authenticatedUser, $suborganization, $data);
 
         if ($invited) {
@@ -476,6 +477,14 @@ class SuborganizationController extends Controller
 
         $data = $suborganizationDeleteUserRequest->validated();
 
+        $userObj = $this->userRepository->find($data['user_id']);
+
+        if ($userObj->hasPermissionTo('organization:view', $suborganization->parent)) {
+            return response([
+                'errors' => ['Can not delete user inherited from a parent org.'],
+            ], 500);
+        }
+
         $is_deleted = $this->organizationRepository->deleteUser($suborganization, $data);
 
         if ($is_deleted) {
@@ -526,6 +535,14 @@ class SuborganizationController extends Controller
         $this->authorize('removeUser', $suborganization);
 
         $data = $suborganizationDeleteUserRequest->validated();
+
+        $userObj = $this->userRepository->find($data['user_id']);
+
+        if ($userObj->hasPermissionTo('organization:view', $suborganization->parent)) {
+            return response([
+                'errors' => ['Can not remove user inherited from a parent org.'],
+            ], 500);
+        }
 
         $authenticatedUser = auth()->user();
